# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sphere English LMS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /quizzes:
    get:
      operationId: getQuizzes
      tags: [quizzes]
      summary: List quizzes
      parameters:
        - name: lessonId
          in: query
          schema:
            type: ["integer", "null"]
        - name: courseId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of quizzes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quiz"

    post:
      operationId: createQuiz
      tags: [quizzes]
      summary: Create a quiz
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuizRequest"
      responses:
        "201":
          description: Quiz created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
  /quizzes/{id}:
    get:
      operationId: getQuiz
      tags: [quizzes]
      summary: Get quiz with questions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quiz details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizDetail"

    patch:
      operationId: updateQuiz
      tags: [quizzes]
      summary: Update quiz
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQuizRequest"
      responses:
        "200":
          description: Updated quiz
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"

    delete:
      operationId: deleteQuiz
      tags: [quizzes]
      summary: Delete quiz
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /quizzes/{id}/submit:
    post:
      operationId: submitQuiz
      tags: [quizzes]
      summary: Submit quiz answers
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizSubmission"
      responses:
        "200":
          description: Quiz result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizResult"
  /quizzes/{id}/results:
    get:
      operationId: getQuizResults
      tags: [quizzes]
      summary: Get quiz results for current user or all students (teacher/admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quiz results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuizResult"

  # Progress
...
  /progress/students/{studentId}:
    get:
      operationId: getStudentProgress
      tags: [progress]
      summary: Get a specific student's progress (teacher/admin)
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgressOverview"
  /progress/courses/{courseId}:
    get:
      operationId: getCourseProgress
      tags: [progress]
      summary: Get progress for all students in a course (teacher/admin)
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course progress for all students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudentCourseProgress"

  # Certificates
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message

    # Auth
...
    Quiz:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        lessonId:
          type: ["integer", "null"]
        courseId:
          type: ["integer", "null"]
        timeLimit:
          type: ["integer", "null"]
        passingScore:
          type: integer
        questionsCount:
          type: integer
        totalPoints:
          type: integer
      required:
        - id
        - title
        - passingScore
        - questionsCount
        - totalPoints
    QuizDetail:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        lessonId:
          type: ["integer", "null"]
        courseId:
          type: ["integer", "null"]
        timeLimit:
          type: ["integer", "null"]
        passingScore:
          type: integer
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
        totalPoints:
          type: integer
      required:
        - id
        - title
        - passingScore
        - questions
        - totalPoints
    Question:
      type: object
      properties:
        id:
          type: integer
        quizId:
          type: integer
        type:
          type: string
          enum: [multiple_choice, true_false, fill_blank, matching, ordering]
        question:
          type: string
        options:
          type: ["array", "null"]
          items:
            type: string
        correctAnswer:
          type: string
        points:
          type: integer
        order:
          type: integer
      required:
        - id
        - quizId
        - type
        - question
        - correctAnswer
        - points
        - order
    CreateQuizRequest:
      type: object
      properties:
        title:
          type: string
        lessonId:
          type: ["integer", "null"]
        courseId:
          type: ["integer", "null"]
        timeLimit:
          type: ["integer", "null"]
        passingScore:
          type: integer
        questions:
          type: array
          items:
            $ref: "#/components/schemas/CreateQuestionRequest"
      required:
        - title
        - passingScore
    CreateQuestionRequest:
      type: object
      properties:
        type:
          type: string
          enum: [multiple_choice, true_false, fill_blank, matching, ordering]
        question:
          type: string
        options:
          type: ["array", "null"]
          items:
            type: string
        correctAnswer:
          type: string
        points:
          type: integer
        order:
          type: ["integer", "null"]
      required:
        - type
        - question
        - correctAnswer
        - points
    UpdateQuizRequest:
      type: object
      properties:
        title:
          type: ["string", "null"]
        timeLimit:
          type: ["integer", "null"]
        passingScore:
          type: ["integer", "null"]
    QuizSubmission:
      type: object
      properties:
        answers:
          type: array
          items:
            $ref: "#/components/schemas/QuizAnswer"
        timeTaken:
          type: ["integer", "null"]
      required:
        - answers
    QuizAnswer:
      type: object
      properties:
        questionId:
          type: integer
        answer:
          type: string
      required:
        - questionId
        - answer
    QuizResult:
      type: object
      properties:
        id:
          type: integer
        quizId:
          type: integer
        quizTitle:
          type: string
        userId:
          type: integer
        userName:
          type: ["string", "null"]
        score:
          type: integer
        totalPoints:
          type: integer
        percentage:
          type: number
        passed:
          type: boolean
        timeTaken:
          type: ["integer", "null"]
        submittedAt:
          type: string
        pointsEarned:
          type: integer
      required:
        - id
        - quizId
        - quizTitle
        - userId
        - score
        - totalPoints
        - percentage
        - passed
        - submittedAt
        - pointsEarned

    # Progress
...
    ProgressOverview:
      type: object
      properties:
        userId:
          type: integer
        totalPoints:
          type: integer
        streak:
          type: integer
        level:
          type: ["string", "null"]
        badges:
          type: array
          items:
            $ref: "#/components/schemas/Badge"
        courseProgress:
          type: array
          items:
            $ref: "#/components/schemas/StudentCourseProgress"
        weeklyActivity:
          type: array
          items:
            $ref: "#/components/schemas/DayActivity"
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/ActivityItem"
      required:
        - userId
        - totalPoints
        - streak
        - badges
        - courseProgress
        - weeklyActivity
        - recentActivity
    Badge:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        earnedAt:
          type: ["string", "null"]
      required:
        - id
        - name
        - description
        - icon
    StudentCourseProgress:
      type: object
      properties:
        userId:
          type: integer
        userName:
          type: ["string", "null"]
        courseId:
          type: integer
        courseTitle:
          type: string
        completedLessons:
          type: integer
        totalLessons:
          type: integer
        percentage:
          type: number
        lastActivity:
          type: ["string", "null"]
      required:
        - userId
        - courseId
        - courseTitle
        - completedLessons
        - totalLessons
        - percentage
    DayActivity:
      type: object
      properties:
        date:
          type: string
        pointsEarned:
          type: integer
        lessonsCompleted:
          type: integer
      required:
        - date
        - pointsEarned
        - lessonsCompleted
    ActivityItem:
      type: object
      properties:
        type:
          type: string
          enum: [lesson_completed, quiz_passed, live_class_attended, badge_earned, course_enrolled]
        description:
          type: string
        pointsEarned:
          type: integer
        timestamp:
          type: string
      required:
        - type
        - description
        - pointsEarned
        - timestamp

    # Certificates
...